WIP memberJoin
