+++ src/components/member/join/button/data/MemberJoinButtons.types.ts
@@ -1,0 +1,6 @@
+import { Member } from '@containers/member/join/member.types';
+import { UseFormReset } from 'react-hook-form';
+
+export type MemberJoinButtonsProps = {
+  reset: UseFormReset<Member>;
+};
