+++ src/components/member/join/button/data/useMemverJoinButtons.tsx
@@ -1,0 +1,8 @@
+import { useTranslation } from 'react-i18next';
+
+const useMemberJoinButtons = () => {
+  const { t } = useTranslation();
+  return { t };
+};
+
+export default useMemberJoinButtons;
