+++ src/components/member/join/button/index.ts
@@ -1,0 +1,1 @@
+export { default as MemberJoinButtons } from './MemberJoinButtons';
