+++ src/components/member/join/index.ts
@@ -1,0 +1,1 @@
+export { default as MemberJoinForm } from './MemberJoinForm';
