+++ src/components/uis/button/InputButton.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const InputButton = () => {
   return <div>InputButton</div>;
 };
